refactor(web): Replace `cheerio` with `@xmldom/xmldom`
Replaced the `cheerio` library with `@xmldom/xmldom` for parsing HTML in the `WebService`.

The primary motivation for this change is to reduce the package's installation size. `cheerio` is a powerful library but come with a significant number of dependencies, increasing the footprint of `ts-coinmarketcap`.

`@xmldom/xmldom` is a much lighter alternative that provides the necessary functionality for parsing the `__NEXT_DATA__` script from the CoinMarketCap website. This change leads to a smaller and more efficient package for end-users without sacrificing functionality.